Reorganizing partner code a little.
